move mint util into utils/testing
