+++ src/pages/DistributePage/useSubmitDistribution.test.tsx
@@ -1,5 +1,5 @@
 import { act, render, waitFor } from '@testing-library/react';
-import { BigNumber, ethers, utils } from 'ethers';
+import { BigNumber, utils } from 'ethers';
 import { GraphQLTypes } from 'lib/gql/__generated__/zeus';
 
 import { useDistributor } from 'hooks';
@@ -13,7 +13,7 @@ import {
   takeSnapshot,
   TestWrapper,
 } from 'utils/testing';
-import { unlockSigner } from 'utils/testing/unlockSigner';
+import { mint } from 'utils/testing/mint';
 
 import {
   SubmitDistributionResult,
@@ -42,65 +42,6 @@ jest.mock('pages/DistributePage/mutations', () => {
     }),
   };
 });
-
-const tokens = {
-  DAI: {
-    addr: '0x6B175474E89094C44Da98b954EedeAC495271d0F',
-    whale: '0x8d6f396d210d385033b348bcae9e4f9ea4e045bd',
-  },
-  USDC: {
-    addr: '0xA0b86991c6218b36c1d19D4a2e9Eb0cE3606eB48',
-    whale: '0x47ac0Fb4F2D84898e4D9E7b4DaB3C24507a6D503',
-  },
-};
-
-async function mint({
-  token,
-  address,
-  amount,
-}: {
-  token: string;
-  address: string;
-  amount: string;
-}) {
-  const mintEth = async (receiver: string, amount: string) => {
-    const signer = provider.getSigner();
-    await signer.sendTransaction({
-      to: receiver,
-      value: ethers.utils.parseEther(amount),
-    });
-    console.log(`Sent ${amount} ETH to ${receiver}`); //eslint-disable-line
-  };
-
-  const mintToken = async (
-    symbol: 'USDC' | 'DAI',
-    receiver: string,
-    amount: string
-  ) => {
-    const { whale, addr } = tokens[symbol];
-    await mintEth(whale, '0.1');
-    const sender = await unlockSigner(whale);
-    const contract = new ethers.Contract(
-      addr,
-      [
-        'function transfer(address,uint)',
-        'function decimals() view returns (uint8)',
-      ],
-      sender
-    );
-    const decimals = await contract.decimals();
-    const wei = BigNumber.from(10).pow(decimals).mul(amount);
-    await contract.transfer(receiver, wei);
-    console.log(`Sent ${amount} ${symbol} to ${receiver}`); // eslint-disable-line no-console
-  };
-
-  switch (token) {
-    case 'USDC':
-    case 'DAI':
-      await mintToken(token, address, amount);
-      break;
-  }
-}
 
 beforeAll(async () => {
   snapshotId = await takeSnapshot();
